chore: move handling of rpc errors to utility
