+++ src/components/AddRpcButton.tsx
@@ -9,6 +9,7 @@ import { darken, transparentize } from "polished";
 import { useConnect } from "@src/hooks/useConnect";
 import { addRpcUrlAction as sendToAnalytics, AddRpcUrlActionType } from "@src/lib/analytics/events";
 import { useWalletClient } from "wagmi";
+import { ERROR_ADD_MANUALLY_MESSAGE, handleRpcError } from "@src/utils/handleRpcError";
 
 
 const Message = styled.p<{ state: AddToWalletStateValues }>`
@@ -37,58 +38,6 @@ const CONNECTING_STATE: AddToWalletState = { state: 'connecting', addOnceConnect
 const TAKING_TOO_LONG_TIME = 15000 // 15s
 const TIMEOUT_TIME = 90000 // 1.5min
 
-const ERROR_USER_REJECTED = { errorMessage: `MEV Blocker was not added. User rejected.`, isUserRejection: true, isError: false }
-const ERROR_ADD_MANUALLY_MESSAGE = { errorMessage: 'There was an error adding the RPC automatically to your wallet. Please add manually', isUserRejection: false, isError: true }
-const ERROR_NETWORK_ADDING_UNSUPPORTED = { errorMessage: `Oh no! 😢 It looks like your wallet doesn't support automatic RPC changes to help protect you. You might be able to make the change manually, though. If you could let your wallet provider know about this, that would be awesome! Thanks for considering it!`, isUserRejection: false, isError: true }
-const ERROR_NETWORK_ALREADY_ADDED = { errorMessage: `Your wallet has a pending request to add the network. Please review your wallet.`, isUserRejection: false, isError: true }
-
-function getErrorMessage(error: any): { errorMessage: string | null, isUserRejection: boolean, isError: boolean } {
-  let actualError = error
-
-  // viem wraps the actual error, we need to get the actual error (not their wrapper)
-
-  if (error.details && typeof error.details === 'string') {
-    try {
-      actualError = JSON.parse(error.details)  
-    } catch {
-      if (error.details === 'Missing or invalid. request() method: wallet_addEthereumChain') {
-        return ERROR_NETWORK_ADDING_UNSUPPORTED
-      }
-    }
-  }
-  
-  const errorCode = actualError?.code
-  if (errorCode === 4001) {
-    return ERROR_USER_REJECTED
-  }
-
-  // -------- Uncomment to debug
-  // if (error) {
-  //   return { errorMessage: JSON.stringify(error, null, 2), isUserRejection: false, isError: true } 
-  // }
-  // -----------------
-
-  const message = error?.details.message || error?.message
-  if (errorCode === -32002 && message?.includes('already pending')) {
-    return ERROR_NETWORK_ALREADY_ADDED
-  }
-
-  if (errorCode === -32000 && (
-    message?.includes('May not specify default') || // i.e. IOS Metamask (over Wallet Connect)
-    message?.includes('Chain ID already exists. Received')) // i.e. Im token
-  ) {
-    return ERROR_NETWORK_ADDING_UNSUPPORTED
-  }
-
-  if (errorCode === -32602 && message?.includes('May not specify default')){
-    // Metakas IOS don't allow you to replace your RPC Endpoint
-    // https://community.metamask.io/t/allow-to-add-switch-between-ethereum-networks-using-api/23595
-    return ERROR_NETWORK_ADDING_UNSUPPORTED
-  }
-
-  return ERROR_ADD_MANUALLY_MESSAGE
-}
-
 export function AddRpcButton() {
   const { addRpcEndpoint } = useAddRpcEndpoint()
   const { isConnected, connect } = useConnect()
@@ -100,7 +49,7 @@ export function AddRpcButton() {
   const [addRpcPromise, setAddRpcPromise] = useState<Promise<boolean> | null>(null)
 
   const handleError = useCallback((error) => {
-    const { errorMessage: message, isError, isUserRejection } = getErrorMessage(error)
+    const { errorMessage: message, isError, isUserRejection } = handleRpcError(error)
     isError && sendToAnalytics('error_add_rpc')
     isUserRejection && sendToAnalytics('rejected_by_user_adding_rpc')
